+++ src/main/java/com/intellij/plugins/haxe/model/type/SpecificTypeReference.java
@@ -402,6 +402,7 @@ public abstract class SpecificTypeReference {
     return null;
   }
 
+  abstract public HaxeResolveResult asResolveResult();
   abstract public SpecificTypeReference withConstantValue(Object constantValue);
 
   final public SpecificTypeReference withoutConstantValue() {